move implementation to mlmodel.py (rebasing)
